move Log to service
