+++ Areas/Admin/Services/ILog.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace PainAssessment.Areas.Admin.Util
+namespace PainAssessment.Areas.Admin.Services
 {
     public interface ILog
     {
